Infer support for this specification mostly
Willy notes that there are plenty of cases where a client might *think* that it
is using early data, but it ends up not being used for a variety of reasons.

This removes the requirement for a client to use `Early-Data`, instead
reserving it for use by intermediaries.  This makes the signal in most cases
implicit: if the handshake isn't done, then it's early.  This complicates the
logic at the server side slightly: the server needs to check the header field
AND the connection status, but I think that's going to be necessary anyway.

This is marginally less conservative, but I think that it retains the spirit of
the original "fix" and it means that most times clients won't have to fiddle
around with the header field.

Closes #6.