Don't repeat TLS advice
TLS mandates the use of anti-replay, but we were implying that it didn't, so
that text needed rewording.

I've moved this to the top of the list, since it is the first line of defense.
I've reworded to emphasize that it is mandatory, but then mention that it is
imperfect.

Closes #39.